Feature EPL-1181: missing properties on interface

+++ src/components/cards/Cards.tsx
@@ -132,7 +132,7 @@ const Cards = ({
   };
 
   const handleDeleteSelectedItems = (itemsToSend: any) => {
-    onDeleteData(itemsToSend);
+    onDeleteData && onDeleteData(itemsToSend);
     handleCancelSelectionMode(false);
   };
